Add the Role entity in the app
Add the Role entity in the app to manage user roles and permissions.
Added to roles: user and advisor, where the default role is user with the lowest level of permissions.
Created seeders for advisors and roles tables.
Run `php artisan migrate --seed` to have roles table populated and advisors user for testing.