Add the Client entity in the app
Client resource holds all information related to a client.
Table 'clients' is related to Client model: email and phone are nullable but at the app level, at least on of them is required.
The ClientSeeder is created for testing purposes to seed 'clients' table.
Relationships between User and Client model have been implemented.