Add the HomeLoanProduct entity in the app
Create 'home_loan_products' table together with the model HomeLoanProduct.
Implement relationships between models. Eager loading is included in Client mode with the protected property `$with`.
NOTE: Added to extra relationships in the User model that can be used later in the app. For now it seems as not needed.
Create seeder for testing purposes HomeLoanProductSeeder.